+++ NEWS.md
@@ -1,5 +1,9 @@
 # Packrat 0.4.9 (unreleased)
 
+- Users can now control whether Packrat snapshots sources using by setting the
+  `R_PACKRAT_SNAPSHOT_SOURCES` environment variable to a 'truthy' value, or by
+  setting the R option `packrat.snapshot.sources`.
+
 - Packrat now provides APIs for accessing the active paths to resource
   directories, with: